refactor: file upload split
